feat: Minimal refactors and new favicon
