+++ lib/dashboard/pages/dashboard_page.dart
@@ -14,22 +14,23 @@ class _DashboardPageState extends State<DashboardPage> {
 
   @override
   Widget build(BuildContext context) {
-    // authService.validateTokenExpiration();
-    // authService.getUser();
-    // final token = authService.getToken();
-    // log(token);
     return const Scaffold(
       body: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
           Icon(
-            Icons.build,
+            Icons.card_giftcard_rounded,
             size: 150,
             color: AppTheme.pinkSalmonShade100,
           ),
           Center(
-            child: Text('Site under construction!',
-                style: TextStyle(fontSize: 150, fontFamily: 'GoldenChesse')),
+            child: Text(
+              'Piki Creativa!',
+              style: TextStyle(
+                fontSize: 150,
+                fontFamily: 'GoldenChesse',
+              ),
+            ),
           ),
         ],
       ),
